pyproject.toml
[tool.poetry]
name = "cms-pdc"
version = "0.1.0"
description = "Gets the PDC stuff"
authors = ["Bentley Hensel <bentleyhensel@gmail.com>"]
license = "AGPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
pyyaml = "^6.0.1"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
